clear sources symlinks directory if cache directory was changed
