+++ lib/config-helpers.js
@@ -1,5 +1,8 @@
 'use strict';
 
+const fs = require('fs-extra');
+const path = require('path');
+const helpers = require('./helpers');
 const ISO639 = {
 
    // Афарский
@@ -563,4 +566,41 @@ function getLanguageByLocale(languageCulture) {
    return ISO639[languageCulture.split('-')[0]];
 }
 
-module.exports = getLanguageByLocale;
+function clearSourcesSymlinksIfNeeded(cacheDirectory) {
+   if (fs.pathExistsSync(path.join(cacheDirectory, 'cache-path.json'))) {
+      const { lastCacheDirectory } = fs.readJsonSync(path.join(cacheDirectory, 'cache-path.json'));
+      if (lastCacheDirectory !== cacheDirectory) {
+         fs.removeSync(path.join(cacheDirectory, 'temp-modules'));
+      }
+   }
+}
+
+function removeLatestSlash(filePath) {
+   if (filePath.endsWith('/') || filePath.endsWith('\\')) {
+      return filePath.slice(0, filePath.length - 1);
+   }
+   return filePath;
+}
+
+/**
+ * проверяем, что результаты компиляции записываются в исходную директорию.
+ * Актуально для мини-задач, например одиночная компиляция typescript или less.
+ */
+function checkForSourcesOutput(config) {
+   const outputDirectory = helpers.unixifyPath(config.output);
+   const { modules } = config;
+   let result = false;
+   modules.forEach((currentModule) => {
+      const moduleDirectory = helpers.unixifyPath(path.dirname(currentModule.path));
+      if (removeLatestSlash(moduleDirectory) === removeLatestSlash(outputDirectory)) {
+         result = true;
+      }
+   });
+   return result;
+}
+
+module.exports = {
+   getLanguageByLocale,
+   clearSourcesSymlinksIfNeeded,
+   checkForSourcesOutput
+};
