+++ lib/helpers.js
@@ -101,30 +101,6 @@ function removeLeadingSlash(filePath) {
    return newFilePath;
 }
 
-function removeLatestSlash(filePath) {
-   if (filePath.endsWith('/') || filePath.endsWith('\\')) {
-      return filePath.slice(0, filePath.length - 1);
-   }
-   return filePath;
-}
-
-/**
- * проверяем, что результаты компиляции записываются в исходную директорию.
- * Актуально для мини-задач, например одиночная компиляция typescript или less.
- */
-function checkForSourcesOutput(config) {
-   const outputDirectory = unixifyPath(config.output);
-   const { modules } = config;
-   let result = false;
-   modules.forEach((currentModule) => {
-      const moduleDirectory = unixifyPath(path.dirname(currentModule.path));
-      if (removeLatestSlash(moduleDirectory) === removeLatestSlash(outputDirectory)) {
-         result = true;
-      }
-   });
-   return result;
-}
-
 const promisifyDeferred = function(deferred) {
    return new Promise((resolve, reject) => {
       deferred
@@ -243,6 +219,5 @@ module.exports = {
    isEqualObjectFirstLevel,
    removeFileFromBuilderMeta,
    joinContents,
-   checkForSourcesOutput,
    componentCantBeParsed
 };
